agents/pyproject.toml
[project]
name = "x-post-agents"
version = "0.1.0"
description = "LangChain agentic content generation for X Post Scheduler"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "pillow>=11.0.0",
    "pydantic-settings>=2.7.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
